+++ contracts/interfaces/IWETH.sol
@@ -3,7 +3,10 @@ pragma solidity ^0.7.0;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address dst, uint wad) external returns (bool);
-    function approve(address guy, uint wad) external returns (bool);
+
+    function transfer(address dst, uint256 wad) external returns (bool);
+
+    function approve(address guy, uint256 wad) external returns (bool);
+
     function balanceOf(address owner) external returns (uint256);
 }
