+++ contracts/uniswapv2/interfaces/IUniswapV2Router.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.7.0;
 
 interface IUniswapV2Router {
-  function factory() external pure returns (address);
+    function factory() external pure returns (address);
 
-  function swapExactETHForTokens(
-    uint amountOutMin, 
-    address[] calldata path, 
-    address to, 
-    uint deadline
-  ) external payable returns (uint[] memory amounts);
+    function swapExactETHForTokens(
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external payable returns (uint256[] memory amounts);
 }
