+++ contracts/utils/FeeHolder.sol
@@ -7,32 +7,33 @@ import "@openzeppelin/contracts/utils/Context.sol";
 /// @title Fee Holder
 /// @author Rafael Romero
 contract FeeHolder is Context, Ownable {
-
     function getBalance() public view onlyOwner returns (uint256) {
         return address(this).balance;
     }
-    
-    function withdraw(uint amount) external onlyOwner {
+
+    function withdraw(uint256 amount) external onlyOwner {
         _withdraw(amount);
     }
-    
-    function transfer(
-        address to, 
-        uint amount
-    ) external onlyOwner returns (bool success){
+
+    function transfer(address to, uint256 amount)
+        external
+        onlyOwner
+        returns (bool success)
+    {
         success = _transfer(to, amount);
         require(success, "FeeHolder: TRANSACTION_FAILED");
     }
-    
+
     function _withdraw(uint256 amount) internal onlyOwner {
         require(address(this).balance >= amount);
         payable(_msgSender()).transfer(amount);
     }
-    
-    function _transfer(
-        address to, 
-        uint256 amount
-    ) internal onlyOwner returns (bool) {
+
+    function _transfer(address to, uint256 amount)
+        internal
+        onlyOwner
+        returns (bool)
+    {
         require(address(this).balance >= amount);
         payable(to).transfer(amount);
         return true;
